// web/src/types/services/expense/ExpenseRequest.ts
export interface ExpenseReadAnyPaginateRequest {
  with_trashed: boolean;
  company_id: string;
  branch_id?: string | null;
  search?: string | null;
  expense_category_id?: string | null;
  is_amount_payable_paid_off?: boolean | null;
  include_id?: string;
  refresh: boolean;
  page: number;
  per_page: number;
}

export interface ExpenseReadAnyGetRequest {
  with_trashed: boolean;
  company_id: string;
  branch_id?: string | null;
  search?: string | null;
  expense_category_id?: string | null;
  is_amount_payable_paid_off?: boolean | null;
  include_id?: string;
  refresh: boolean;
  limit: number;
}